+++ src/events/ready.ts
@@ -1,5 +1,3 @@
-import type { Client, ClientPresence } from 'discord.js';
-import { ActivityType } from 'discord.js';
 import { Worker } from 'node:worker_threads';
 
 import { appVersion } from '../constants/meta.js';
@@ -38,10 +36,6 @@ export const ready = onEvent('ready', {
 		info('Verifying command deployments...');
 		await verifyCommandDeployments(client);
 
-		// Set user activity
-		info('Setting user activity');
-		setActivity(client);
-
 		// Start uptime ping
 		const UPTIME_URL = process.env['UPTIME_URL'];
 		if (UPTIME_URL) {
@@ -55,11 +49,3 @@ export const ready = onEvent('ready', {
 		info('Ready!');
 	},
 });
-
-function setActivity(client: Client<true>): ClientPresence {
-	// Let users know where to go for info
-	return client.user.setActivity({
-		type: ActivityType.Playing,
-		name: '/help for info',
-	});
-}