+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix Prisma type errors from `npm run setup` script
+- Set user activity on Client initialization, keeps presence from disappearing as much
 
 ## [0.15.0] - 2025-09-25